pyproject.toml
[tool.poetry]
name = "async-libxduauth"
version = "0.1.0"
description = ""
authors = ["Blindinlights <chenyutong007@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.5"
crypto = "^1.4.1"
bs4 = "^0.0.1"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.1"
pycryptodome = "^3.18.0"
lxml = "^4.9.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
